Add api guards to routes. Update auth type. update utils is auth function

+++ src/app/api/cats/[id]/route.ts
@@ -1,7 +1,9 @@
 import { eq } from "drizzle-orm";
+import { type NextRequest, NextResponse } from "next/server";
 import { db } from "~/server/db";
 import { cats } from "~/server/db/schema";
 import type { ICat } from "~/types";
+import { isAdmin } from "../../utils";
 
 /** api/cats/[id] */
 
@@ -22,7 +24,11 @@ export async function PATCH(request: Request, { params }: { params: { id: number
   return Response.json(updatedRecord);
 }
 
-export async function DELETE(_request: Request, { params }: { params: { id: number } }) {
+export async function DELETE(request: NextRequest, { params }: { params: { id: number } }) {
+  if(!(await isAdmin(request))){
+    return NextResponse.json({message: "Forbidden"}, {status: 403});
+  }
+
   const { id } = params;
   const [deletedRecord] = await db.delete(cats)
     .where(eq(cats.id, id))
